# Code/Server/docs/swagger.yaml
basePath: /v1
definitions:
...
  db.Role:
    enum:
    - owner
    - tenant
    type: string
    x-enum-varnames:
    - RoleOwner
    - RoleTenant
...
  models.IdResponse:
    properties:
      id:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      profile_picture_url:
        type: string
      role:
        $ref: '#/definitions/db.Role'
      updated_at:
        type: string
    type: object
  models.UserUpdateRequest:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
    type: object
...
  utils.Error:
    properties:
      code:
        $ref: '#/definitions/utils.ErrorCode'
      error:
        type: string
    type: object
  utils.ErrorCode:
    enum:
    - invalid-password
    - cannot-fetch-user
    - user-not-found
    - cannot-create-user
    - no-claims
    - cannot-decode-user
    - missing-fields
    - missing-file
    - cannot-hash-password
    - email-already-exists
    - test-error
    - too-many-requests
    - invite-not-found
    - user-must-have-same-email-as-invite
    - invite-already-exists-for-email-or-property
    - lease-already-exists-for-tenant-and-property
    - property-not-found
    - property-is-not-yours
    - not-an-owner
    - not-a-tenant
    - user-already-exists-as-owner
    - property-already-exists
    - property-not-available
    - tenant-already-has-lease
    - no-active-lease
    - lease-not-found
    - cannot-end-non-current-lease
    - inventory-report-must-be-linked-to-current-lease
    - no-pending-lease
    - document-not-found
    - damage-not-found
    - damage-already-exists
    - cannot-update-fixed-damage
    - damage-already-fixed
    - failed-upload-file
    - failed-delete-file
    - bad-base64-string
    - property-picture-not-found
    - user-profile-picture-not-found
    - room-not-found
    - room-already-exists
    - room-not-in-this-property
    - furniture-not-found
    - furniture-already-exists
    - furniture-not-in-this-property
    - inventory-report-already-exists
    - cannot-modify-inventory-report
    - inventory-report-not-found
    - room-state-already-exists
    - furniture-state-already-exists
    - invalid-object-type
    - error-request-chatgpt-api
    - failed-send-email
    type: string
    x-enum-varnames:
    - InvalidPassword
    - CannotFetchUser
    - UserNotFound
    - CannotCreateUser
    - NoClaims
    - CannotDecodeUser
    - MissingFields
    - MissingFile
    - CannotHashPassword
    - EmailAlreadyExists
    - TestError
    - TooManyRequests
    - InviteNotFound
    - UserSameEmailAsInvite
    - InviteAlreadyExists
    - LeaseAlreadyExist
    - PropertyNotFound
    - PropertyNotYours
    - NotAnOwner
    - NotATenant
    - UserAlreadyExistsAsOwner
    - PropertyAlreadyExists
    - PropertyNotAvailable
    - TenantAlreadyHasLease
    - NoActiveLease
    - LeaseNotFound
    - CannotEndNonCurrentLease
    - InvReportMustBeCurrentLease
    - NoLeaseInvite
    - DocumentNotFound
    - DamageNotFound
    - DamageAlreadyExists
    - CannotUpdateFixedDamage
    - DamageAlreadyFixed
    - FailedUploadFile
    - FailedDeleteFile
    - BadBase64String
    - PropertyPictureNotFound
    - UserProfilePictureNotFound
    - RoomNotFound
    - RoomAlreadyExists
    - RoomNotInThisProperty
    - FurnitureNotFound
    - FurnitureAlreadyExists
    - FurnitureNotInThisProperty
    - InventoryReportAlreadyExists
    - CannotModifyInventoryReport
    - InventoryReportNotFound
    - RoomStateAlreadyExists
    - FurnitureStateAlreadyExists
    - InvalidObjectType
    - ErrorRequestChatGPTAPI
    - FailedSendEmail
host: localhost:3001
info:
  contact:
    email: martin.d-herouville@epitech.eu
    name: Mazettt
  description: This is the API used by the Immotep application.
  termsOfService: http://swagger.io/terms/
  title: Immotep API
  version: "1.0"
paths:
...
  /profile/:
    get:
      consumes:
      - application/json
      description: Get current user profile
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get current user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update current user profile
      parameters:
      - description: User update info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.Error'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Update current user profile
      tags:
      - user
  /profile/picture/:
    put:
      consumes:
      - multipart/form-data
      description: Update current user's profile picture
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile picture
        in: formData
        name: picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Updated user ID
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Missing fields or bad base64 string
          schema:
            $ref: '#/definitions/utils.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Update current user's profile picture
      tags:
      - user
...
  /users/{id}/:
    get:
      consumes:
      - application/json
      description: Get user information by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"